separate gateway client and server
